Configuration is now multi-machine: graphite+sonar
Changed the configuration to have two machines started so I can run
graphite in one isolated machine, and a "sonar" machine in the other.

The sonar declaration is only for ease of separation; its not going to run
sonar itself, but simply by used as a quick source box to graphite.

Changes include

 * graphite config pushed into a config element
 * sonar configuration created
 * networking changed from public to private to create a mini-cluster